requirements.txt
streamlit
pymupdf
numpy
pillow
google-generativeai